requirements.txt
flask
PyJWT
google-auth-oauthlib
google-auth
pylint
pytest 
gunicorn
selenium
dotenv
mysql-connector-python
opensearch-py
pytest-cov
psycopg2-binary
certbot
python-dotenv
boto3>=1.28.0,<2.0.0
urllib3>=1.25.4,<2.0.0